+++ tests/test_delineate.py
@@ -5,7 +5,9 @@ import os
 import shutil
 import unittest
 from subprocess import check_output
+
 from basin_setup.basin_setup import *
+
 from .basin_setup_test_case import BSTestCase
 
 
@@ -20,9 +22,9 @@ class TestDelineateCLI(BSTestCase):
         self.pour_points = join(self.data_path, 'pour_points.bna')
         self.dem = join(self.data_path, 'dem_epsg_32611_100m.tif')
         self.cmd_str = 'delineate -d {} -p {}  -o {}'.format(
-                                                    self.dem,
-                                                    self.pour_points,
-                                                    self.output
+            self.dem,
+            self.pour_points,
+            self.output
         )
 
     def test_ensemble(self):